Cleanup server side renderer
This makes server part of webpack build way faster during development,
~10x.

Benchmark (my xps):
- OLD: 19.35s
- NEW:  1.97s
